Audio/requirements.txt
absl-py==2.3.1
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
alembic==1.17.2
antlr4-python3-runtime==4.9.3
anyio==4.11.0
asteroid-filterbanks==0.4.0
async-timeout==5.0.1
attrs==25.4.0
audioread==3.1.0
av==15.1.0
backports.cached-property==1.0.2
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
coloredlogs==15.0.1
colorlog==6.10.1
contourpy==1.3.2
ctranslate2==4.6.1
cycler==0.12.1
decorator==5.2.1
docopt==0.6.2
einops==0.3.2
exceptiongroup==1.3.1
faster-whisper==1.2.1
ffmpeg-python==0.2.0
filelock==3.20.0
flatbuffers==25.9.23
fonttools==4.60.1
frozenlist==1.8.0
fsspec==2025.12.0
future==1.0.0
gmpy2 @ file:///C:/b/abs_d8ki0o0h97/croot/gmpy2_1738085498525/work
greenlet==3.2.4
grpcio==1.76.0
h11==0.16.0
hf-xet==1.2.0
hmmlearn==0.2.8
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.15.1
humanfriendly==10.0
HyperPyYAML==1.2.2
idna==3.11
Jinja2==3.1.6
jiwer==4.0.0
joblib==1.5.2
julius==0.2.7
kiwisolver==1.4.9
lazy_loader==0.4
librosa==0.11.0
lightning==2.5.6
lightning-utilities==0.15.2
llvmlite==0.45.1
Mako==1.3.10
Markdown==3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.7
mdurl==0.1.2
mkl-service==2.5.2
mkl_fft @ file:///C:/miniconda3/conda-bld/mkl_fft_1761592981169/work
mkl_random @ file:///C:/miniconda3/conda-bld/mkl_random_1761593075610/work
more-itertools==10.8.0
mpmath==1.3.0
msgpack==1.1.2
multidict==6.7.0
networkx==3.4.2
nltk==3.9.2
numba==0.62.1
numpy==1.26.4
omegaconf==2.3.0
onnxruntime==1.23.2
openai-whisper==20250625
optuna==4.6.0
packaging==25.0
pandas==2.2.1
pillow==12.0.0
platformdirs==4.5.0
pooch==1.8.2
primePy==1.3
propcache==0.4.1
protobuf==3.20.1
pyannote.audio==2.1.1
pyannote.core==4.5
pyannote.database==4.1.3
pyannote.metrics==3.2.1
pyannote.pipeline==2.3
pycparser==2.23
pyDeprecate==0.3.2
Pygments==2.19.2
pyparsing==3.2.5
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytorch-lightning==1.9.0
pytorch-metric-learning==1.7.3
pytz==2025.2
PyYAML @ file:///C:/miniconda3/conda-bld/pyyaml_1763465513320/work
RapidFuzz==3.14.3
regex==2025.11.3
requests==2.32.5
rich==14.2.0
ruamel.yaml==0.18.16
ruamel.yaml.clib==0.2.15
safetensors==0.7.0
scikit-learn==1.7.2
scipy==1.15.3
semver==2.13.0
sentence-transformers==2.2.2
sentencepiece==0.2.1
shellingham==1.5.4
simplejson==3.20.2
singledispatchmethod==1.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soundfile==0.13.1
soxr==1.0.0
speechbrain==0.5.16
SQLAlchemy==2.0.44
sympy==1.14.0
tabulate==0.9.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.4
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers==0.13.3
tomli==2.3.0
torch==2.2.2
torch-audiomentations==0.12.0
torch_pitch_shift==1.2.5
torchaudio==2.2.2
torchmetrics==0.11.4
torchvision==0.17.2
tqdm==4.67.1
transformers==4.30.2
typer==0.20.0
typer-slim==0.20.0
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
Werkzeug==3.1.3
whisperx @ git+https://github.com/m-bain/whisperx.git@d32ec3e3012ec4c0934f4088424c32f3f038b249
yarl==1.22.0
